refactor(kilo,opencode): split handler.ts into handler modules
Both backends wrap the same upstream HTTP API + SSE event stream, so
they share a parallel handler/ layout:
  - state   — in-flight session registry (getActiveSession)
  - turn    — runXTurn + SSE subscription + last-assistant lookup
  - message — handleMessage entry point + post-loop accounting + the
              formatEventCounts diagnostic

kilo/handler.ts was 718 lines, opencode/handler.ts was 677. index.ts in
each re-exports the same public surface; importers updated to the barrel.
